refactor(front): uniformized components
